Fix timezones in recurring reminder confirmation responses
The bot would say "OK, I will remind you in 1 minute, and every 1 week and 1 hour" even if you
specified 1 week intervals, as timezones are a thing and suck.

Just subtract the utcoffset retrieved from the timezone to get the actual timedelta the
user wanted.

Also fix in the !listreminders print-out.